requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
scikit-learn
opencv-python
seaborn
jupyter
json
Pillow
tensorflow-addons